Adding django database that uses sqlite
you can check out the models.py to see how the db was implemented and you can go to /admin to see and play around with all the tables/data. everything in the templates is just placeholders. also rn the app is just called 'website' but obviously we can change that