database - made task and task component inside of fronchettineese room
feature - refactored the room editor and task so now the task has a list of all its components and a function that is called by room editor to serialize all components. This was in order to make the task component creation process easy

ux - added additional error for login if incorrect credentials provided or if api is not connectinf ot backend

feature - changed rturn of auth api to contain message and status code allowing for proper message on frontend. Plan to implement for uture apis